added F1,F2,F3 descriptors for a total of 7660
due to the existence of 5 pairs instead of 1, F_i includes about 2500 descriptors instead of the 72 from the original paper. these are not included in X class descriptors, so I should be able to still just be able to manage creating them as well. I estimate about 175,000 descriptors in this class, and since creating the current 7660 itakes about 10 seconds, I optimistically expect this to take 4-5 minutes